+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ES: no more whiste screen of death when VRAM is filled
 - sound: only set volume for playback devices
 - psx: fixed typo (mu3 instead of m3u file extension)
+- Moonlight: bump to 2.4.6, supports up to GFE 3.12 + can handle several GFE PC + rewrote scrape
+- recalbox.conf videomode now supports a new auto mode
 
 ## [18.02.09] - 2018-02-09
 - C2: disable picodrive for now until we fix compilation